Fix resizing issues.
The old resizing code made it impossible to actually resize the window.
The current code is not perfect, but at least it allows setting the size
of the window, and the window gets resized automatically when the video
size changes.

Note: It doesn't get back the same size when the video size gets back to
normal because we don't handle changes in the requisitionned size of the
panel (which can change when the displayed text changes).

svn path=/trunk/; revision=7186