Potentially fixed problems with uris and party names.
OPAL has a weird API to determine the remote uri and party names.

The remote party name defaults to the remote party address when it is
unknown. Try guessing it is the case and use a sane default in that case
instead of the OPAL one.

Added more special chars to drop from OPAL uris by default.