Support to send DTMFs from anywhere in the main GUI.
Formerly, when you were in call and wanted to send a DTMF, it was
necessary to switch to dialpad and press the appropriate button.

It is now possible to press any allowed key from anywhere in the main
window when it has the focus. That way, the user does not have to
constantly switch between the dialpad and the roster when he wants
to input DTMFs to a remote system.