Fixed crashes shown as assertion failed in /usr/include/ptlib/safecoll.h:813
The fix is the same as 43c6c4, but in other places in the source code.

The crash was 100% reproducible: Start ekiga, wait until it registers.
Add in the URI echo@ekiga.net, call, you receive an error.  Go with
cursor in the URI, remove echo, type 500, but immediately after typing
5, there is crash.