Make sure the output device is set before a call
Fixes the following bug: The main thread was closing the audioinput
device, while the "Media Patch" thread which pumps the data into the
internet was still active.  Consequently "Media Patch" switched to the
silence device.  This was left in current_device of
audioinput-core.cpp, and it was used for the next call.  Attached
patch also sets internal_set_manager(desired_device); when the stream
is started, i.e. when it is really needed.