Fixed a few crashes related to clicking on contacts/presentities
Those are due to the fact that the reference counts aren't
embedded anymore in the objects pointed to by smart pointers,
so when we put a raw pointer in a gtk+ tree store, we must
be careful not to put it back in a smart pointer which will
kill it prematurely. Likewise, doing shared_ptr<Foo>(this)
leads to a dead end unless one uses a null deleter.