Fixed bug #594433.
It is now possible to register dsandras@ekiga.net as user name to ipness.net.
This is another form of SIP registration similar to a proxy. The correct way
to work would be to use a proxy, but some providers do not document things that
way leading to user confusion.