[windows] Fix major bug where the uninstaller removes non-Ekiga files
This case appeared when Ekiga is installed in a non-empty directory,
because the uninstaller removes the whole install directory, hence
other files too.

The simplest solution I found is that if the installation directory is
not empty (such as MyDocs), then it is installed into an empty
sub-directory.  Thus, even when the uninstall removes the whole
sub-directory, there is no problem.