Add workaround for Busy Here bug
When a call is received while we are in communication, from then on
ekiga will reply Busy Here to all incoming calls.

This commit is just a hack, waiting a proper fix (in opal or ekiga).

This commit also introduces another smaller bug: if you receive a call
while in a call, Ekiga will no longer show Missed call.