Turned the call history view code into a real GObject with signals
This fixes bug #630792.

I thought I would only have to add the signals, but it wasn't even a
GObject! It did require some surgery in both the initialization and the
destruction code : we have to setup a GObject, and deal with disposing
(disconnecting signals for both boost and gtk+) in the right function.

It compiles and survives rapid tests, so it must not be that bad.