Removed ekiga_main_window_selected_build_menu
I don't get why it was called in ekiga_main_window_update_calling_state,
so I removed it... and once I removed it, it was used only once.

Then I moved its body in place... and saw it could simplify the code by
removing the 'presentity' member.

I saw a few possible bugs, so I'll move on and fix them.