The contact menu in the main window is now updated when something happens
Previously, once the contact was selected, the menu was just there. If the
contact menu changed, we wouldn't change ; and on removal, we waited until
it disappeared from the view (which finally changed the selection) to act.

Now, we check both for updates and for removal, so the menu in the GUI is
synchronized with the menu of the presentity.