Made the contact menu also work for the selected call history item
It is updated when the item is updated or removed, it is greyed when moving
away -- unless of course you go in the roster view with a selected
presentity, in which case you get that presentity's menu. Come back and you
get the history item menu back.

It compiles and seems to just work.