Make the main window handle shutdown better
Its code is neither gtk+ code nor engine code... so it lacks a proper
dispose method : when the roster view gets killed, it updates its
selection ; unfortunately the main window is already half-dead but doesn't
know it yet. Sigh... I added a check on what we need : is it still there?