Made the OnPresenceChange for the opal presence callback work better by simplifying the code
It was making strange computations to decide what to do... now it just makes
easy translations (from an enum to strings) and pushes the information around.

The code can probably still be simplified, and made more complete with respect to the enum,
but it's still a big improvement : I think opal presence mostly works!