Fixed leaking the opal account store in the assistant
In fact, the problem was that the private structure of that GObject was created using pure C code...
That means the C++ ctor/dtor of the members weren't being called. I made that structure managed using
new/delete, and now things look good.

Notice that we were probably leaking the video input core, the audio input core and the audio output core too!