Presence: Only fetch presence for current account.
We were fetching presence of the same list of uris for each account
instance. Only fetch if account is enabled.

This asks the question: shouldn't we fetch presence only if the uri
correspond to the account at the presence core level instead of passing
the list of all uris to all account instances?