Accounts: Fixed account toggling/untoggling and presence.
When you enable an account, it now subscribe to presentities again.
We just update watched_uris even if we do not fetch presence info
because the corresponding account is not enabled.

When you disable an account, the presentity information changes
to unknown so that wrong information is not displayed.

A bug remains in opal though when subscribe/unsubscribe/subscribe.