Main Window: Moved audio and video statistics.
They are now displayed in the call panel instead of the status bar.

This makes sense especially if we want to support several calls at the
same time. The stats would be call specific in that case.

Moreover, the statusbar should be used only for status messages like
missed calls and so on.

For now, we limit the global notifications to incoming calls. This
should change soon.