UI: Cosmetic changes.
The URI bar moved to the call panel: people should start calls from
instant messaging or from the contacts list. If they are more advanced
users, they will enter uris directly, and that will be done in the call
panel. This leaves more space to the contacts list.

The call panel will be moved out of the main window. It will be
presented as incoming call popup on incoming calls. If the contacts list
is hidden, it will stay hidden.

The advantage is more usability and better engine usage : the call panel
will contain a notebook with one page per call. One video widget, but
several call control areas.

This is the next commit (on Wednesday).