Main Window: Implemented new actions toolbar.
The general UI design was also improved.

This new actions toolbar provides a shortcut to the various panel
sections of the main window. This is more convenient to use than
switching between pages using the menu or the notebook arrays. It also
provides a more compact user interface.

I made use of GNOME icons in order to make them themeable.

The video preview button was moved out of the call window back into the
main window as suggested by Thierry Simonet. This provides the video
preview action back.