Default action in the Call Menu is now Contact based.
I have moved the "Call" action right below the contact related
actions because our UI is contact centric.

This displays the dialpad.

If we wanted to refine things, we should either call the selected
contact (if any), or display the dialpad (if no contact selected).
However, it seems rather difficult to implement and will be done later.