Notifications: Better notifications.
I'm not satisfied yet. If we use gnome-shell, probably we should have
notifications associated to Ekiga's icon and packing up.

However, this is not the way libnotify works right now, so probably we
should just live with it.

As a consequence, there is no clear way for users to have an history of
notifications and we are forced to keep the status icon for that.