Libnotify: Moved incoming call notifications to the engine.
We have a nice libnotify component displaying notifications coming from
various parts of the engine core.

Incoming call notifications require a bit more knowledge of what happens
with the call and need to be able to act on it. Nevertheless, the engine
libnotify component is the right place for this.