Incoming call dialog: Dropped it.
Dropped incoming call dialog and replace it by the call window.

The call window is now automatically shown on incoming calls if our
notification server does not support actions.
This test is done through a local static method.

An answer button was added to the dialog.

If the notification server supports actions, we rely on the engine to
display the incoming call notification and handle it.