Chat: Improved notifications.
Added notifications triggered by the ChatWindow itself when there are
unread text messages. The ChatWindow is indeed the only component that
knows if a message has been read or not.

The notification(s) allow displaying the chat window with the unread
messages.

Also dropped the notify.* files. They are not necessary anymore. All
notifications are handled by the NotificationCore.