Notifications: Added Audio*Device detection.
Changing the audio device when a new device is added or a
device removal happens is now done through Notifications.

Handle the case where the added/removed device is
the prefered one or not.

This allows the main window to get rid of those signals.

We do not support changing the 'Ringer' device anymore. However, it did
not really made sense as it is supposed to be the soundcard (which is
hardwired in the laptop) in most cases.