Statusicon: Do not rely on notifications persistence to show it.
The problem is that GNOME 3 Fallback also supports persistence, but has
a notification area.

Now, if gnome-shell is running, the status icon is not shown. Closing
the window hides it. We expect the user to launch ekiga again or to wait
for some action to show it back.

If gnome-shell is not running, we display the statusicon in the status
area and we expect him to click on it to show it back.