GUI: Improved accounts window.
The checkbox was misleading: it could not be toggled/untoggled.

It has now been replaced by a nice icon indicating the account status.
The icon is synchronised with the published user status. That way, the
user rapidly knows which accounts are active or not, and what is the
published status for those accounts.