h323: set bandwidth according to network type
The bandwidth for the H323 endpoint was fixed to 40000 (4 Mbps) without any
apparent reason. Meanwhile, in Opal, the default is 100000 (10 Mbps).

It is clear then that the H323 bandwidth should be associated with the
configured network type and not a fixed value.

This patch assigns the bandwidth mapped from the configured network type:

PSTN & ISDN = 128 kbps
DSL = 4 Mbps
LAN & default = 10 Mbps

This mapping is also arbitrary, but it makes more sense than just a fixed
value.