common-videooutput: store the extended stream buffer
When a second remote video stream (extended stream) arrives, it is
stored in a new byte array called eframeStore. The dimensions of
the frame are also stored, along with a new boolean property that
confirms the existence of an extended stream in the connection.

This patch does not modify the current user experience, but it is a
requirement for the extended video display.