common-videooutput: add VO_MODE_REMOTE_EXT mode
This patch adds the extended remote video stream enum in the video output
mode.

Also it adds the switch labels where the enum is handled in order to avoid
compiler's complains, but no logic is added in those cases, in other words it
is just the skeleton. In the following commits that logic will be added.