videoutput: new parameter device_opened() signal
This patch is traversal to all the videooutput subsystem.

It adds a new parameter to the on_device_opened() signal. This new parameter
indicates if there is an extended stream in the connection.

With it, the UI can enable the option to display the extended stream in the
video widget.