ui: add the extended video option
With this patch the user can choose to display the "extended video" through
the view menu.

The menu item is enabled only when the on_videoutput_device_opened_cb()
callback has the extended stream (ext_stream) as true.

When the opened video output device doesn't contain an extended stream, and
the view is set on that stream, the configuration is reset to the local view.

Also, the validation of the values for the key
/apps/ekiga/general/user_interface/video_display/video_view is more generic,
using enum symbols rather than numbers.