videooutput: new parameter size_changed() signal
This patch is traversal to all the videooutput subsystem.

It adds a new parameter to the size_changed() signal. This new parameter
indicates the video output mode of the current frame that changed its size.

With this new parameter the UI can know which window resize in case the
output is other window.

https://bugzilla.gnome.org/show_bug.cgi?id=687775