Notifications: Fixed bug where notification disappaers too fast.
It seems that when you set the transient hint before the notification is
shown, it is ignored. The transient hint is set for incoming call
notifications only.

If you set it after the notification has been shown, then the notification
is really transient : the persistence of the notification daemon is ignored
and the notification stays until you accept or reject the call, or until
there is a timeout.