Opal: Fixed call auto-answer.
It seems the way we were doing things triggered a bug in the new
OPAL code. It has been fixed and we are now doing things properly.

Alert-Info based auto-answer has been removed for 4.0. It could lead to
any incoming call to be answered and not just the one we wanted.
Moreover, an option should be present to allow or disallow it.

Otherwise, anybody could trigger an auto-answered call to any Ekiga
user, which would not be very nice.