ekiga.schemas: fix a possible UDP port collision
The SIP listen port is configured to be 5060, and the default UDP port
range is from 5060 to 5100. This means that the first UDP port used by
Opal is going to be 5060, but it will already be used by the SIP
endpoint, causing an error in the connection. At the next port
binding, the problem will be solved "magically".

This issue has raised in the gatekeeper registration process: after
setting up the SIP and the H323 endpoints, the H323 endpoint starts
the gatekeeper registration. Opal uses the first port in the UDP range
to connect to the configured gatekeeper, i.e. 5060, but the
registration silently fails, because the port is already in use. The
workaround was to unregister and register again.

It remains to do the migration on current ekiga installations.

Closes bug #690621.