Roster: Fixed bug #692370.
All sorts of crashes could happen if the presentity icon was updated by
the blinking source timer after it had been deleted.

We now remove the timer when a presentity is destroyed.

I suppose it could also happen when quitting too faster after having
started the application, but I am not sure about this.