[windows] Fix freeze on quit
This commits replaces delete x (where x inherits PThread) with calling
a quit function which does the same thing as the destructor, and
letting ptlib deleting the thread later.

Before this commit, the thread calling delete always frozen on
TerminateThread function.  Anyway, calling this function is dangerous
(see
http://msdn.microsoft.com/en-us/library/windows/desktop/ms686717%28v=vs.85%29.aspx),
so we must ensure that the target thread has finished all its job,
which is what this commit does.

Fixes bug #671818.