X: remove the graphic context handling
The main reason behind to keep depending on GTK+-2.0 is because the
rendering subsystem demanded the graphic context of the widget. But this is not
mandatory. You can create a graphic context and use it in your widget.

This patch removes the graphic context from the XWindow and XVWindow
constructors and creates a new one in their initializations. Now Ekiga
is not tied to GTK+-2.0, at least because of this.

This behavior is similar what GStreamer's xvimagesink does.

See https://bugzilla.gnome.org/show_bug.cgi?id=685404

Thanks to Víctor Manuel Jáquez Leal for this patch!