GSettings: Do not remember positions for the prefs and assistant.
Those were never remembered. The problem with the assistant, for
example, is that it derivates from GtkWindow, and it is impossible to
make it derivate from our GmWindow object. Remembering position and size
is probably too difficult for those.

Please note that it is the WM job to remember those settings. However,
we have our own code for non-Linux ports.