Added an experimental Opal::Presentity implementation
This is heavily based on Local::Presentity's code, but with some little
improvements here and there (pushing build_node as a static method makes
a better api, for instance), removed presence fetching/unfetching code
(which will need to be re-added later, perhaps differently).

It hasn't been added to the build system, and was never compiled.