Fixed flickering effect when windows are shown.
This was due to the fact the window was first shown, then moved to the
last saved position. This seems to be "new" with GTK3. We can now move
the window to the correct position when the realize signal is emitted.

That prevents the flickering effect. However, it requires the GmWindow
to be initialized from the start with the correct GSettings key. The
call window and the main window were the only one with a differed key
setup. This is now fixed too.

The former gm_window_set_key methods have been removed as they can not
be used anymore.