AudioInputCore: Added device auto-selection algorithm.
The responsability to choose the best device for the user is now left to
the Core. The Core will select the best default device in case the
choosen setting is invalid.

On Linux, we respectively prefer PulseAudio, then ALSA, then the first
available device if the others are not found.