Audio/Video Cores: Added anoter get_devices method.
The new method returns a vector of devices represented by their string
descriptions.

The string description is used by the prefs and in various parts of the
program. The new method has the big advantage of having the same
signature whatever the core is unlike the former method which required
different arguments for each core.