Clutter GST: Added a new clutter gst videooutput.
This is using the GStreamer and Clutter APIs to display frames.

The new videooutput designs requires the videooutput to display frames
onto a surface whatever its size, position and states are. That's
implemented by the current code which will output frames on a
ClutterActor. The part of the GUI managing the ClutterActor is
responsible to size it appropriately following the natural size of the
video.