Ekiga: Started the GtkApplication port.
The GtkFrontend class was removed. The old file now contains a new
GtkApplication (GmApplication) GObject responsible of the whole
application.

Currently, the responsability is limited to creating Ekiga application
windows and displaying the toplevel menu in GNOME Shell. (Other menus
have not been ported to the new model, so you might be confronted to
uncomplete menus if you are not using GNOME Shell at all. Please be
patient!

The new GmApplication holds a shared_ptr to the Ekiga::ServiceCore.

As part of the move, the various windows API has been reorganized to:
 - Only take the GmApplication as argument
 - Register as a GtkApplicationWindow of the main GtkApplication

Some code with FIXME has also been moved when it had to be moved.

Most windows are now following the new "create" and "destroy" model
instead of being kept in memory. The only window where it was not
possible is the chat window. That's for later.

The global gmcallbacks have also been removed and integrated into
GtkApplication.