UI: Added call action to main window + UI fixes.
This adds a call button to the main window UI. Actions being
added/removed on the fly depending on the context (e.g. if a contact is
clicked), the button will become sensitive/unsensitive automatically
depending on the Action availability.

This is especially useful for global actions like "call".

The patch also fixes buttons alignment to be more conform to what we
would expect from GNOME programs.