Engine+OPAL: Added new global actions.
Global actions were added to implement the minimum functions we would
expect from a softphone:
- Call (coming from the CallCore)
- Transfer (coming from the CallCore)
- IM (coming from the CallCore, but should probably be moved in the
  ChatCore)

The corresponding methods were added to the Core and to the respective
CallManager/CallProtocolManager classes because they are the minimal
subset of operations we want endpoints to implement. More specific
actions will be exposed to the Actor mechanism too, but will not be part
of the Core API.

The old populate_menu stuff has been (partially) removed.