Actors: Added ContactActor. Added the notion of disabled Actions.
This largely simplifies the ActorMenu code.

The ContactActor class handles ContactActions. The set_data method is
used to set data on all ContactActions of the ContactActor.

The Actor object now can enable/disable Actions by name. This will allow
disabling Actions on the fly after they have been added. Disabled
actions will not appear in menus.

When data is set on a ContactAction, the ContactAction is automatically
disabled when it can not be run with the given data.

This allows having an oversimplistic ActorMenu/ContactActorMenu
implementation. The ContactActorMenu methods have mostly been removed as
everything can be handled in the ActorMenu parent class thanks to the
notion if "enabled" Actions.