Actor: Added a new and simpler Action/Actor mechanism.
This allows exposing Actions to the UI when they are not specifically
part of the generic API.

An object able to execute Actions is an Actor. An Actor is exposing its
own Actions which can also be pulled from an ActionProvider (e.g. the
CallCore could provide call Actions to a Contact).