Forms: Fixed flickering effect on error.
The new Forms implementation allows callbacks to fail by returning
false, and to place an error message in the arguments.

This permits our GTK+ implementation to display the error in the current
form and to accept Form validation only in case there is no error.

The former implementation was creating a new form instance displaying
the various fields, their values and the error message. That new popup
replacing the validation form was triggering an annoying virtual
artifact.