Forms: Added allow_empty parameter to Text Form properties.
It allows defining Forms where some Text entries can not be empty. The
user will be prevented form validating the Form until the corresponding
Form text entries have been filled with some text.

It becomes unnecessary to do the check in the callback (and to report an
error in case of problem).

It has been implemented because HIG3 suggests preventing the user to
validate incorrect data (when possible).