GmEntry: Added a activate_icon property and an activated signal.
The activate icon is an icon that appears when the GmEntry content is
both valid and non-empty (even if allow-empty is set to true). It emits
the activate signal when being clicked. (this is the reason why it can
not be empty).

The activated signal is emitted when the GtkEntry activate signal is
emitted and the GmEntry content is valid (and not empty).