GmInfoBar: Add the ability to display several messages.
It often happens that the application generates several messages that
the user needs to acknowledge. The previous implementation was able to
display one message at a time, overriding older messages without
allowing the user to acknowledge or read them. This is now fixed by the
implementation of a stack of messages.