Actions: Fixed behavior to be conform with the rest of Ekiga.
When a contact is selected in the roster, the actions_changed signal is
emitted with all possible Contact and Heap actions.

This is common to all Ekiga subcomponents: actions_changed always
carries all available Actions for the selected item (including parent
actions). However, right-click menus only present the specific actions
for the selected item.