CodecDescription: Large rework of CodecDescription and CodecList.
The general idea behind this change is to move codec list loading from
the configuration where it belongs: in the child Opal class and not in
the main base class. This allows more control over what we are doing and
different CallManager could have different codec handling mechanisms.
The newly introduced load method allows loading codecs descriptions from
the configuration is dependant on the child implementation. The Opal
component uses its own "load" method.

The CodecList class has been simplified. Instead of handling an internal
list, it inherits of a std::list of std::string, where std::string is
the internal representation of the enabled/disabled codec.

The codecs blacklist is now handled in the Opal CodecList
implementation. This is cleaner.

Furthermore, we do not rely on the SDP encoding name for the codec. This
allows the user to change the priority between several variations of the
same codecs, e.g. H.264 High-Profile and H.264.

Finally, a known-codecs.h file has been added. Even though codecs are
plugins, it allows giving more useful information to the user.