CallCore: Added methods to get and set codecs.
This will use the different CallManagers to do so.

It allows the CodecsBox to rely on the CallCore instead of the
configuration keys content to get a codec description.

This offers more flexibility and the possibility to handle codecs in
different CallManagers in a uniform way wrt the corresponding
CallManager feature.