SIP: Fixed incorrect account state notifications.
Sometimes, a failed registration was notified as a successful
unregistration.

We should only handle registration state transitions we are interested
in.  For example, if we were registering, we are interested in knowing
if the registration was successful or not. If the registration was not
successful and OPAL cleanly unregisters the account (as a
consequence), we are not interested in knowing the
unregistration worked.