Opal: Sanitized Services and shared_ptr usage.
This is the first step towards complete cleaning of the OPAL
implementation regarding Services and shared_ptr.

The H.323 and SIP EndPoints are not services anymore. There is no point
in having them added to the Services core.

Their respective objects are now passed as references to the Opal Bank
and Accounts which are using them instead of using share_ptr's. This
prevents having cyclic references, most notably when objects should be
destroyed. Accounts should be destroyed before Banks, which should in
turn be destroyed before the EndPoints. shared_ptr's are useless in this
can as we control the objects lifetime.