Opal: Reorganized more code.
The SIP & H323 Opal EndPoints/CallProtocolManagers creation has been
moved from opal-main.cpp to the OpalCallManager, where it belongs.

The Opal Bank & Accounts now only rely on the OpalCallManager and not on
individual CallProtocolManagers. The Opal Account object now directly
accesses the Opal Endpoints to register/unregister accounts. This code
can actually be self-contained into the Opal objects and not rely on
engine objects. That allows simplifying object lifecycles.

Endpoints are now "setup" when the CallManager signals it is ready. That
means we do not need direct references to the SIP & H.323 endpoints into
the CallManager like it was the case.