BankImpl: Removed wrong API comment.
Any deleted BankImpl is automatically removed from the AccountCore.

This is not true anymore, and it is probably bad practice in this case.
We are using shared_ptr's, which means that the underlying object will
be deleted only when there are no more references to that object (if by
miracle, there are no cyclic references).

If we wait for the object to be deleted to remove it from the
AccountCore, it will never be removed as the AccountCore is a list (and
not a RefLister?) of shared_ptr's.

We need to work the other way around : each object using a Bank/BankImpl
needs to be destroyed and stop using it before the object is actually
automatically destroyed.