Opal: Fix process so that it takes care of destroying what it created.
The Opal Process created a Bank that was added as a PresencePublisher
and as a Cluster to the PresenceCore, and as a Bank to the AccountCore.

When the Opal Process exits, it should take care of removing that Bank
from those specific containers. That way, at least the Opal Process does
not keep references it added.