Presence: Make sure the Ekiga::Cluster are destroyed properly on quit.
We were binding shared_ptr to the Cluster objects signals.

That means that the shared_ptr are garbage collected
at a later stage, at least with boost::signals2, which is not what we
want. One solution would be to use weak_ptr's. However, we are not
really using the Cluster share_ptr in any of our callbacks, so
simplifying the signals signature seems to be a better idea.

Signals relay was also simplified, useless callbacks have been removed.