Opal: Started a big refactoring.
This commit is still incomplete. Some features do not work yet, and some
methods are not where they should be.

The idea behind this refactoring is that we now have a separate OPAL
process, with separate H.323 and SIP endpoints. They are running in
standalone mode.

However, the engine requires CallManagers in order to place calls. Those
engine SIP and H.323 CallManagers will use the relevant OPAL process to
implement actions.

The reason behind this change is that both parts (the engine and the
opal part) manage memory differently. Moreover, our code was becoming
really big with a single endpoint implementing various parts of the
engine. Things are now more modular and easier to maintain and
understand.

The CallProtocolManager was removed as part of this change!