Opal: Reorganized process destruction.
Shutting down endpoints and cleaning up factories are executed by Opal
when the process is being destroyed. However, it triggers segfaults when
using static PMutex instances in OpalGloballyUniqueID for example.

That is why we clean things up preventively.

An engine_close method has been added. It allows manual cleaning when
required for some parts of the engine.