Opal: Let the Engine handle the Call destruction.
Boost will delete it when convenient.

Of course, it introduces a leak due to the fact that we use things of
the form:
  objects[obj]->add (obj->updated.connect (boost::bind (boost::ref
  (object_updated), obj)));

It means that our objects binds on itself with a shared_ptr to itself as
argument. The new boost garbage collector will prevent the object to be
destroyed.