DynamicObject: Added replacement for RefLister and LiveObject.
The new replacement is working around the crossed references problem
preventing boost to release memory.

The problem is due to signal relay: we connect signals of an object so
that they are relayed by another object using the first object as
argument.

The solution is to pass the object itself to the signal instead of
binding to it.