Opal: Officially drop STUN.
STUN is not a good way of handling NAT problems. Modern clients should
either rely on ICE (that we do not support yet) or rely on providers
that support NAT traversal transparently for the client.

Early versions of STUN have been deprecated.

We will move Ekiga.net to support this.