Opal: Simplified call end reasons.
We had too many cases for various different subtilities users are not
interested in. They simply should know if the call has been interrupted
in a normal way or because of an error: call denied, abnormal call
termination, ...

In traditional telephony, we only have 3 failure states: busy, invalid
number and congestion. We should stick to that.