Roster: Fixed offline presentities display.
We should only consider presentities as offline when we know they are
offline through presence. If we have no presence information at all, we
should not consider them as offline and we should not hide them.

They could be cellphone numbers or traditional numbers added to the
roster.

Or the PBX could simply not support that information.