audiooutput: fallback to primary device if secondary is not set
Ekiga already falls back to the primary audio output device
when the manager for the secondary device is NULL. But the
gmconf bridge will set the SILENT device if an empty setting for
the secondary device is encountered, which effectively mutes
ekiga's ring tone unless the user actively reconfigures the
secondary device. The primary device, if not set, defaults to
the first detected audio device on the system, which is a
more reasonable default than SILENT.

This patch changes the behavior for the secondary device such
that the primary device fallback is used for the secondary
device if no configuration is found.

Solves: https://bugs.launchpad.net/ubuntu/+source/ekiga/+bug/558304